backend/requirements.txt
aiosqlite==0.17.0
anyio==3.5.0
asgiref==3.5.1
astroid==2.11.5
black==22.3.0
click==8.1.3
dill==0.3.4
fastapi==0.77.1
flake8==4.0.1
greenlet==1.1.2
h11==0.13.0
httptools==0.4.0
idna==3.3
iso8601==1.0.2
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
parso==0.8.3
pathspec==0.9.0
platformdirs==2.5.2
pycodestyle==2.8.0
pydantic==1.9.0
pyflakes==2.4.0
pylint==2.13.8
pypika-tortoise==0.1.4
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.36
starlette==0.19.1
tomli==2.0.1
tortoise-orm==0.19.0
typing_extensions==4.2.0
uvicorn==0.17.6
uvloop==0.16.0
watchgod==0.8.2
websockets==10.3
wrapt==1.14.1
